chore: Center the plus icon so I don't die
This took me way too long to figure out, but the plus icon was off-center with respect to its viewbox. I hand-edited the path to center it on 8,8... how it should be. To see the difference, just pull up the production deployment side-by-side with the preview deployment and inspect the plus icon in the "new chat" button.